% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{wndr_file}
\alias{wndr_create_file}
\alias{wndr_delete_file}
\alias{wndr_file}
\alias{wndr_get_file}
\title{File API}
\usage{
wndr_get_file(id = NULL, task_id = NULL, list_id = NULL)

wndr_create_file(task_id, filepath, local_created_at = NULL)

wndr_delete_file(id, revision)
}
\arguments{
\item{id}{File ID}

\item{task_id}{Task ID}

\item{list_id}{List ID}

\item{filepath}{Path to a local file you want to upload.}

\item{local_created_at}{Date used to sort the files and comments in the right order}

\item{revision}{Revision}
}
\description{
A task may contain files.
}
\details{
When you use \code{wndr_get_file}, you have to specify either one of \code{id}, \code{task_id}, \code{list_id}.
}
\examples{
\dontrun{
l <- wndr_get_list()

wndr_get_file(list_id = l$id[1])

t <- wndr_get_task(list_id = l$id[1])
f <- wndr_create_file(task_id = t$id[1], filepath = "/path/to/file")

wndr_delete_file(id = f$id, revision = f$revision)
}

}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/file}
}

