% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{wndr_folder}
\alias{wndr_create_folder}
\alias{wndr_delete_folder}
\alias{wndr_folder}
\alias{wndr_get_folder}
\alias{wndr_update_folder}
\title{Folder API}
\usage{
wndr_get_folder(id = NULL)

wndr_create_folder(title, list_ids)

wndr_update_folder(id, revision, title = NULL, list_ids = NULL)

wndr_delete_folder(id, revision)
}
\arguments{
\item{id}{Folder ID}

\item{title}{Folder name}

\item{list_ids}{List IDs}

\item{revision}{Revision}
}
\description{
A list may or may not belong to a folder.
}
\examples{
\dontrun{
# get all folders
wndr_get_folder()

# get a folder
wndr_get_folder(id = 1111)

# create a folder
l <- wndr_get_list()
f <- wndr_create_folder(title = "test", list_ids = l$id[1:3])

# update the folder
wndr_update_folder(id = f$id, revision = f$revision, list_ids = l$id[2:3])

# delete the folder
wndr_delete_folder(id = f$id, revision = f$revision)
}

}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/folder}
}

