% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avatar.R
\name{wndr_get_avatar}
\alias{wndr_get_avatar}
\title{Avatar API}
\usage{
wndr_get_avatar(user_id, size, fallback = FALSE)
}
\value{
matrix, the parsed result of response PNG image.
}
\description{
Avatar API
}
\examples{
\dontrun{
  x <- wndr_get_avatar(1, 128, fallback = TRUE)
  plot(as.raster(x))
}
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/avatar}
}

