% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{wndr_get_file}
\alias{wndr_get_file}
\title{File API}
\usage{
wndr_get_file(id = NULL, task_id = NULL, list_id = NULL)
}
\arguments{
\item{id}{File ID}

\item{task_id}{Task ID}

\item{list_id}{List ID}
}
\description{
A task may contain files.
}
\details{
When you use \code{wndr_get_file}, you have to specify either one of \code{id}, \code{task_id}, \code{list_id}.
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/file}
}

