% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{wndr_get_folder}
\alias{wndr_get_folder}
\title{Folder API}
\usage{
wndr_get_folder(id = NULL)
}
\description{
Folder API
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/folder}
}

