% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{wndr_get_list}
\alias{wndr_get_list}
\title{List API}
\usage{
wndr_get_list(id = NULL)
}
\arguments{
\item{id}{List id}

\item{title}{List title}

\item{revision}{list revision}
}
\description{
List API
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/list}
}

