% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{wndr_get_list_position}
\alias{wndr_get_list_position}
\title{Positions API}
\usage{
wndr_get_list_position(id = NULL)
}
\description{
Positions API
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/positions}
}

