% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{wndr_get_note}
\alias{wndr_get_note}
\title{Note API}
\usage{
wndr_get_note(id = NULL, task_id = NULL, list_id = NULL)
}
\description{
Note API
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/note}
}

