% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{wndr_get_preview}
\alias{wndr_get_preview}
\title{File Previews}
\usage{
wndr_get_preview(file_id, platform = NULL, size = NULL)
}
\description{
File Previews
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/file_preview}
}

