% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtask.R
\name{wndr_get_subtask}
\alias{wndr_get_subtask}
\title{Subtask API}
\usage{
wndr_get_subtask(id = NULL, task_id = NULL, list_id = NULL,
  completed = NULL)
}
\description{
Subtask API
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/subtask}
}

