% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{wndr_get_task}
\alias{wndr_get_task}
\title{Task API}
\usage{
wndr_get_task(id = NULL, list_id = NULL, completed = NULL)
}
\description{
Task API
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/task}
}

