% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{wndr_get_user}
\alias{wndr_get_user}
\title{User API}
\usage{
wndr_get_user(list_id = NULL)
}
\arguments{
\item{list_id}{List ID}
}
\description{
All info related to the currently signed in user.
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/user}
}

