% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhook.R
\name{wndr_get_webhook}
\alias{wndr_get_webhook}
\title{Webhook API}
\usage{
wndr_get_webhook(list_id)
}
\description{
Webhook API
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/webhooks}
}

