% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{wndr_list}
\alias{wndr_create_list}
\alias{wndr_delete_list}
\alias{wndr_get_list}
\alias{wndr_list}
\alias{wndr_update_list}
\title{List API}
\usage{
wndr_get_list(id = NULL)

wndr_create_list(title)

wndr_update_list(id, revision, title = NULL, public = NULL)

wndr_delete_list(id, revision)
}
\arguments{
\item{id}{List id}

\item{title}{List title}

\item{revision}{List revision}

\item{public}{Whether or not the list is public}
}
\description{
All tasks created in Wunderlist belong to a list.
}
\examples{
\dontrun{
# get all lists
wndr_get_list()

# get a list
wndr_get_list(id = 1111)

# create a list
l <- wndr_create_list("test")

# update the list
wndr_update_list(id = l$id, revision = l$revision, title = "Test")

# make the list public
wndr_update_list(id = l$id, revision = l$revision, public = TRUE)

# delete the list
l <- wndr_get_list(id = l$id)
wndr_delete_list(id = l$id, revision = l$revision + 2)
}

}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/list}
}

