% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{wndr_membership}
\alias{wndr_create_membership}
\alias{wndr_delete_membership}
\alias{wndr_get_membership}
\alias{wndr_membership}
\alias{wndr_update_membership}
\title{Membership API}
\usage{
wndr_get_membership()

wndr_create_membership(list_id, user_id = NULL, email = NULL,
  muted = NULL)

wndr_update_membership(id, revision, state = "accepted", muted = NULL)

wndr_delete_membership(id, revision)
}
\arguments{
\item{list_id}{List ID}

\item{user_id}{User ID}

\item{email}{User email address}

\item{muted}{Whether or not muted}

\item{id}{Membership ID}

\item{revision}{Revision}

\item{state}{Membership state. If set this to \code{"accepted"}, the membership is accepted}
}
\description{
A Membership is the join model between Users and Lists.
}
\examples{
\dontrun{
# get memberships
wndr_get_membership()

# create a membership
l <- wndr_get_list()
m <- wndr_create_membership(list_id = l$id[2], user_id = 111111)

# mark it as accepted
wndr_update_membership(id = m$id, revision = m$revision, state = "accepted")

# delete the membership
wndr_delete_membership(id = m$id, revision = m$revision + 1)
}

}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/membership}
}

