% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{wndr_note}
\alias{wndr_create_note}
\alias{wndr_delete_note}
\alias{wndr_get_note}
\alias{wndr_note}
\alias{wndr_update_note}
\title{Note API}
\usage{
wndr_get_note(id = NULL, task_id = NULL, list_id = NULL)

wndr_create_note(task_id, content)

wndr_update_note(id, revision, content)

wndr_delete_note(id, revision)
}
\arguments{
\item{id}{Note ID}

\item{task_id}{Task ID}

\item{list_id}{List ID}

\item{content}{Note content}

\item{revision}{Revision}
}
\description{
A note is a large text blobs belonging to tasks.
}
\details{
When you use \code{wndr_get_note}, you have to specify either one of \code{id}, \code{task_id}, \code{list_id}.
}
\examples{
\dontrun{
# get notes
l <- wndr_get_list()
wndr_get_note(list_id = l$id[2])

# create a note
t <- wndr_get_task(list_id = l$id[2])
n <- wndr_create_note(task_id = t$id[1], content = "test")

# update the note
wndr_update_note(id = n$id, revision = n$revision, content = "Test")

# delete the note
n <- wndr_get_note(id = n$id)
wndr_delete_note(id = n$id, revision = n$revision)
}

}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/note}
}

