% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{wndr_preview}
\alias{wndr_get_preview}
\alias{wndr_preview}
\title{File Previews}
\usage{
wndr_get_preview(file_id, platform = NULL, size = NULL)
}
\arguments{
\item{file_id}{File ID.}

\item{platform}{Platform. Possible values are: mac, web, windows, iphone, ipad, android.}

\item{size}{Size of screen. Possible values are: nonretina, retina}
}
\description{
A file preview is an image thumbnail for a file.
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/file_preview}
}

