% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminder.R
\name{wndr_reminder}
\alias{wndr_create_reminder}
\alias{wndr_delete_reminder}
\alias{wndr_get_reminder}
\alias{wndr_reminder}
\alias{wndr_update_reminder}
\title{Reminder API}
\usage{
wndr_get_reminder(task_id = NULL, list_id = NULL)

wndr_create_reminder(task_id, date, created_by_device_udid = NULL)

wndr_update_reminder(id, revision, date)

wndr_delete_reminder(id, revision)
}
\arguments{
\item{task_id}{Task ID}

\item{list_id}{List ID}

\item{date}{Timestamp when to remind.}

\item{created_by_device_udid}{Device UDID, which is used for push notification.}

\item{id}{Reminder ID}

\item{revision}{Revision}
}
\description{
The reminder for a task.
}
\examples{
\dontrun{
# get reminders
l <- wndr_get_list()
r <- wndr_get_reminder(list_id = l$id[1])

# create a reminder
t <- wndr_get_task(list_id = l$id[1])
r <- wndr_create_reminder(task_id = t$id[1], date = Sys.time() + 3600)

# update the reminder
r <- wndr_update_reminder(id = r$id, revision = r$revision, date = Sys.time() + 3600 * 24)

# delete the reminder
wndr_delete_reminder(id = r$id, revision = r$revision)
}

}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/reminder}
}

