% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{wndr_root}
\alias{wndr_get_root}
\alias{wndr_root}
\title{Root API}
\usage{
wndr_get_root()
}
\description{
Root is the top-level entity in the sync hierarchy.
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/root}
}

