% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtask.R
\name{wndr_subtask}
\alias{wndr_create_subtask}
\alias{wndr_delete_subtask}
\alias{wndr_get_subtask}
\alias{wndr_subtask}
\alias{wndr_update_subtask}
\title{Subtask API}
\usage{
wndr_get_subtask(id = NULL, task_id = NULL, list_id = NULL,
  completed = NULL)

wndr_create_subtask(task_id, title, completed = NULL)

wndr_update_subtask(id, revision, title = NULL, completed = NULL)

wndr_delete_subtask(id, revision)
}
\arguments{
\item{id}{Subtask ID}

\item{task_id}{Task ID}

\item{list_id}{List ID}

\item{completed}{Whether the subtask is completed or not}

\item{title}{Subtask title}

\item{revision}{Revision}
}
\description{
Subtasks are children of tasks.
}
\examples{
\dontrun{
# get subtasks
l <- wndr_get_list()
s <- wndr_get_subtask(list_id = l$id[1])

# create a subtask
t <- wndr_get_task(list_id = l$id[1])
s <- wndr_create_subtask(task_id = t$id[1], title = "test")

# update the subtask
s <- wndr_update_subtask(id = s$id, revision = s$revision, completed = TRUE)

# Delete the subtask
wndr_delete_subtask(id = s$id, revision = s$revision)
}

}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/subtask}
}

