% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{wndr_task}
\alias{wndr_create_task}
\alias{wndr_delete_task}
\alias{wndr_get_task}
\alias{wndr_task}
\alias{wndr_update_task}
\title{Task API}
\usage{
wndr_get_task(id = NULL, list_id = NULL, completed = NULL)

wndr_create_task(list_id, title, assignee_id = NULL, completed = FALSE,
  recurrence_type = c("day", "week", "month", "year"),
  recurrence_count = NULL, due_date = NULL, starred = NULL)

wndr_update_task(id, revision, title = NULL, assignee_id = NULL,
  completed = NULL, recurrence_type = c("day", "week", "month", "year"),
  recurrence_count = NULL, due_date = NULL, starred = NULL,
  remove = NULL)

wndr_delete_task(id, revision)
}
\arguments{
\item{id}{Task ID}

\item{list_id}{List ID.}

\item{completed}{Whether the task is completed or not.}

\item{title}{Task title.}

\item{assignee_id}{User ID.}

\item{recurrence_type}{Type of reccurence. Possible values are: "day", "week", "month", "year".}

\item{due_date}{Due date.}

\item{starred}{Whether the task is starred or not.}

\item{revision}{Revision.}

\item{remove}{Array of attributes to delete from the task, e.g. \code{"due_date"}.}

\item{reccurence_count}{Count of reccurence.}
}
\description{
Tasks are children of lists.
}
\examples{
\dontrun{
# get tasks
l <- wndr_get_list()
t <- wndr_get_task(list_id = l$id[1])

# create a task
t <- wndr_create_task(list_id = l$id[2], title = "test")

# create a recurrence task
wndr_create_task(list_id = l$id[2], title = "test", recurrence_type = "week", recurrence_count = 1, due_date = Sys.Date())

# update the task
t <- wndr_update_task(id = t$id, revision = t$revision, completed = TRUE)

# delete the task
wndr_delete_task(id = t$id, revision = t$revision)
}

}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/task}
}

