% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_comment.R
\name{wndr_task_comment}
\alias{wndr_create_task_comment}
\alias{wndr_get_task_comment}
\alias{wndr_task_comment}
\title{Task Comment API}
\usage{
wndr_get_task_comment(id = NULL, task_id = NULL, list_id = NULL)

wndr_create_task_comment(task_id, text)
}
\arguments{
\item{id}{Comment ID}

\item{task_id}{Task ID}

\item{list_id}{List ID}

\item{text}{Text}
}
\description{
A task_comment is a comment that belongs to a task.
}
\examples{
\dontrun{
# get task comments
l <- wndr_get_list()
c <- wndr_get_task_comment(list_id = l$id[1])

# create a task comment
t <- wndr_get_task(list_id = l$id[1])
wndr_create_task_comment(task_id = t$id[1], text = "test")
}

}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/task_comment}
}

