% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{wndr_upload}
\alias{wndr_upload}
\title{Upload API}
\usage{
wndr_upload(filepath)
}
\arguments{
\item{filepath}{Path to a file}
}
\description{
Upload a file to use it via File API.
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/upload}
}

