% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhook.R
\name{wndr_webhook}
\alias{wndr_create_webhook}
\alias{wndr_delete_webhook}
\alias{wndr_get_webhook}
\alias{wndr_webhook}
\title{Webhook API}
\usage{
wndr_get_webhook(list_id)

wndr_create_webhook(list_id, url, processor_type = "generic",
  configuration = "")

wndr_delete_webhook(id, revision)
}
\arguments{
\item{list_id}{List ID}

\item{url}{Webhook URL}

\item{processor_type}{Type of processor.}

\item{configuration}{Configuration}

\item{id}{Webhook ID}

\item{revision}{Revision}
}
\description{
A webhook sends notifications when a list is updated.
}
\seealso{
\url{https://developer.wunderlist.com/documentation/endpoints/webhooks}
}

